+++ frontend/screens/meetings/CreateMeetingScreen.js
@@ -2,7 +2,7 @@ import * as React from "react";
 import { Alert, View } from "react-native";
 import { Button, TextInput } from "react-native-paper";
 import { getNickname } from "../../UserManager";
-import { createMeeting } from "../../Api";
+import { createMeeting } from "../../api/MeetingsApi";
 
 export default function CreateMeetingScreen({ navigation, route }) {
   const userUuid = route.params.userUuid;
